ViewRootImpl: Don't pause HWUI so much
We have paused and thus flushed the HW renderer before calling relayout for many releases.

This has served a multitude of purposes over the years:
1. Ensuring that the WM doesn't destroy the surface or update the
   default buffer size while we are rendering - WM no longer updates
   default buffer size or handles surface lifetime
2. Ensure that we can get a proper frame number, previously used for
   deferTransactionUntil - deferTransactionUntil doesn't exist anymore
3. Ensure that we can call setNextTransaction at an appropriate time
   - We now call this from an RT thread side callback.

This means in the current code we only have to pause it when we ourselves are updating the size. Since this pause requires all previous draws to flush, it should be a nice performance win.

Test: Existing tests pass
Bug: 220649859
Change-Id: I143de3f44b8bd69754d9e824fc8b729dc296e183