Apply ringer mode user settings to notification vibrations
The vibrator service is only applying the ringer mode settings to
ringtone vibrations, forcing them to be disabled when ringer mode is
silent.

This is applied to notification vibrations by the notification manager,
but some other paths (including 3p apps) might apply this usage and the
ringer mode settings would not be applied to them (e.g. Dialer and
Telecom app/service, Car services, etc).

This also applies the flag FLAG_BYPASS_INTERRUPTION_POLICY on ringer
mode, since the CellBroadcastAlert can trigger alerts with the
notification usage and those need to bypass user settings and vibrate
all the time.

Bug: 198794501
Test: VibrationSettingsTest
Change-Id: Ic00fbd32d859b0c77830450add24826e0dd13366