Remove ImsConfigTracker#onBindInput precondition
It is expected the onIntialize() is called before onBindInput(), however
it's not precisely known why onBindInput() is being called prior to onIntialize().
Once possibility is it could be that onDestory() is being called
and init was skipped and call made it to onBindInput().

Replacing precondition with an early return should be safe here.

Fix: 223083664
Bug: 223443508
Test: atest CtsInputMethodTestCases

Change-Id: I8312d98b0349fbc661b6eb0f4c47bea25ffbb2a9