Bouncer: Remove overlap with shade and bouncer
Scale values of fading away shade content. Also scale the notification
scrim alpha as well when state is SHADE_LOCKED.

There are a few ternaries to preserve the state of the notification
shade when device is unlocked. (The notification shade content will not
disapper as quickly if the bouncer is not about to show).

There also seems to be some flickering of the notification scrim alpha
when enter bouncer from the notification shade and successfully passing
the security method, but this seems to be a separate issue.

Bug: 226108150
Test: Manual on device
Change-Id: Id717191cefe3901983ef6a46174383110d89f5ae