Detect "expand to fullscreen" over recents and switch to it
If a pip (or any other window) is expanded to fullscreen, it will
occlude recents. In this situation, we should immediately finish
the recents "animation" and let the new one play so we don't hang.

This also adds a new "callback" for "replace with screenshot".
We need this because if we just finish recents, it will immediately
destroy the "live" tile. So this callback lets us replace the
"live" tile with a screenshot beforehand.

Bug: 223321653
Test: Put app in Pip, open another app, go to recents, expand pip
Change-Id: I8a5ca05d5eef43e37c74f38b7988003bad7cdade