Move logging to background thread in LogBuffer
To avoid checking log level settings on the main thread, which can cause jank, move the settings check and logging to a background thread. During manually testing with both the original mechanism and the background thread, messages still appeared in the console with the exact same timestamp (to the millisecond). Nothing I did on the phone was able to get the background thread's BlockingQueue to ever hold more than 3 messages at any given time.

Bug: 221419865
Test: Manually verified messages still logged and BlockingQueue had sufficient capacity
Change-Id: If230e09674dcf1dd7e7b1026e71717f982c5b3d2